funding: added funding field to deposit.
