+++ src/lib/components/Funding/NoAwardResults.js
@@ -1,0 +1,41 @@
+// This file is part of React-Invenio-Deposit
+// Copyright (C) 2021 CERN.
+// Copyright (C) 2021 Northwestern University.
+//
+// React-Invenio-Deposit is free software; you can redistribute it and/or modify it
+// under the terms of the MIT License; see LICENSE file for more details.
+
+import PropTypes from 'prop-types';
+import React from 'react';
+import { Message } from 'semantic-ui-react';
+import { i18next } from '@translations/i18next';
+
+export function NoAwardResults({ switchToCustom }) {
+    console.log('Rendering NoAwardResults');
+    console.log(switchToCustom);
+  return (
+    <Message
+      warning
+      icon="warning circle"
+      header={i18next.t('No results found')}
+      content={
+        <p>
+          {i18next.t('No awards matching your criteria were found.')}
+          <a
+            href="#"
+            onClick={(e) => {
+              e.preventDefault();
+              switchToCustom();
+            }}
+          >
+            {i18next.t('Add a custom award.')}
+          </a>
+        </p>
+      }
+    />
+  );
+}
+
+NoAwardResults.propTypes = {
+  switchToCustom: PropTypes.func.isRequired,
+};
